+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
-**3.0.16 - 10/30/24**
+**3.0.16 - 10/31/24**
 
   - Bugfix to prevent a LookupTable from changing order of the value columns
+  - Fix mypy errors in vivarium/framework/lookup/table.py
+  - Typing changes in vivarium/framework/lookup/interpolation.py
+  - Fix broken build from LayeredConfigTree typing
 
 **3.0.15 - 10/24/24**